moved GetDepositAccountBalance
